+++ src/api.js
@@ -1,4 +1,5 @@
-const rocketsApi = 'https://api.spacexdata.com/v3/rockets';
+const rocketsApi = 'https://api.spacexdata.com/v3';
+const ROCKETS = '/rockets';
 
-export const rockets = () => fetch(rocketsApi).then((res) => res.json());
+export const rockets = () => fetch(rocketsApi + ROCKETS).then((res) => res.json());
 export const nonrockets = 1;
